+++ earth2studio/lexicon/__init__.py
@@ -34,6 +34,7 @@ from .metop import (
     MetOpMHSLexicon,
 )
 from .mrms import MRMSLexicon
+from .mtg import MetOpMTGLexicon
 from .ncar import NCAR_ERA5Lexicon
 from .nclimgrid import NClimGridLexicon
 from .planetary_computer import (